Remove Xerces from OpenSim, fix CMakeLists accordingly and add test case targeted to old version files. All test cases pass in release mode except for testNCSpline which was failing before the merge.
git-svn-id: file:///dev/shm/opensim/Trunk@6754 fc29be9c-2cb0-420d-b520-73dbd18add6e

+++ OpenSim/Simulation/Wrap/WrapCylinderObst.cpp
@@ -287,7 +287,7 @@ int WrapCylinderObst::wrapLine(const SimTK::State& s, SimTK::Vec3& aPoint1, SimT
 	// Generate wrap_pts sequence of points tracing out wrapping path
 	aWrapResult.wrap_pts.append(aWrapResult.r1);
 	double Qang=atan2(Qy,Qx);				// Angle of point Q
-	int i, n=1+(QtoTang>=0?QtoTang:-QtoTang)/max_wrap_pts_circle_ang;	// Number of angle steps from Q to T angles
+	int i, n=1+ (int)((QtoTang>=0?QtoTang:-QtoTang)/max_wrap_pts_circle_ang);	// Number of angle steps from Q to T angles
 	double angDelt=(QtoTang)/(double)(n);	// Delta angle for n steps from Q to T angles
 	for(i=0;i<=n;i++) {
 		double ang = Qang + i*angDelt;		// Angle ranging from that of Q to that of T
